Fixed #016523: Viewing survey results doesn't work in oracle
When doing an automatic install that uses the .dba file to import the database schema,
longtext fields were created as CLOB on Oracle. This caused a problem on ezsurveyquestionresult.text
as this field is used in a GROUP BY clause to view results, and this is incompatible with CLOBs.
A special configuration has been added to dbschema.ini in order to have it converted to VARCHAR2(4000).
An update script can be found in update/database/oracle/2.3/